+++ web_assets.go
@@ -184,11 +184,12 @@ load();pollTimer=setInterval(load,750);
 `
 
 const defaultConfigJSON = `{
-  "Version": 9,
-  "SavedBy": "8.3.0-go-default",
+  "Version": 10,
+  "SavedBy": "8.4.0-rc3-go-default",
   "SavedAt": "2026-07-05T00:00:00+09:00",
   "ActiveProfileId": "default",
   "AutoSwitch": {"Enabled": false, "DebounceMs": 350, "Bindings": []},
+  "Controller": {"Enabled": false},
   "Profiles": [
     {
       "Id": "default",